requirements.txt
appnope==0.1.3
asttokens==2.0.8
backcall==0.2.0
biopython==1.79
certifi==2022.6.15
charset-normalizer==2.1.1
decorator==5.1.1
executing==1.0.0
idna==3.3
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.6
numpy==1.23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
stack-data==0.5.0
tdqm==0.0.1
tqdm==4.64.0
traitlets==5.3.0
urllib3==1.26.12
wcwidth==0.2.5
